Support prefixes for partial template names
If we want to use this more than once within an app - such as when a module wants to add its own partial templates to the stack - then to avoid conflicts with similarly named files we can add an optional prefix to namespace the partial names without needing to force the file structure.

Also export partials with absolute rather than relative paths to facilitate use of multiple 'views' directories.